Add files via upload
Added changes to navigation bar like getting rid of bullet points, blue colouring and lines under the links and spaced the words out. I made a space around the links so you can press on a box around it. Next steps are to make the homepage and sub-pages look good.