End of Version two!
In this version I added different background photos and text for the cards on every sub-page. I changed the title and description for the sub-pages and made the background images different to make the different pages somewhat unique. They have the same format which could make it a bit boring which is why I made these changes to keep it interesting.