Version three done
In this version I moved the navigation bar to the left because my stakeholder decided they liked it better on the left and added a hover effect which smoothly underlines the text on the nav bar. This will be my final addition to the website, concluding the developing phase of this project.